Create FindMiddleOfLinkedList.cpp
Link :  https://www.geeksforgeeks.org/dsa/split-a-circular-linked-list-into-two-halves/
Traverse the circular linked list using the fast and slow pointer technique to find the middle and last nodes. The slow pointer will reach the middle, and the fast pointer will reach the end.
If the list has an odd number of nodes, the fast pointer will reach the last node. If the list has an even number, it will stop just before the last node.